# apps/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bankstract-cloud API",
    "description": "Statement parsing API for Nigerian banks. Parse a PDF/XLSX statement into structured transactions over a single HTTP call. API consumers interact over HTTP and do not inherit the AGPL-3.0 license of the hosted service.",
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/banks": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "List supported banks",
        "operationId": "banks_v1_banks_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BanksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Current billing-period usage",
        "operationId": "usage_v1_usage_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/keys": {
      "post": {
        "tags": [
          "Keys"
        ],
        "summary": "Issue an API key",
        "description": "Admin-only. Returns the raw key exactly once; only its argon2 hash is stored.",
        "operationId": "create_key_v1_keys_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyCreatedResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Keys"
        ],
        "summary": "List API keys",
        "operationId": "list_keys_v1_keys_get",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/keys/{key_id}": {
      "delete": {
        "tags": [
          "Keys"
        ],
        "summary": "Revoke an API key",
        "operationId": "revoke_key_v1_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/parse": {
      "post": {
        "tags": [
          "Parse"
        ],
        "summary": "Parse a bank statement",
        "description": "Upload a PDF (or XLSX) statement and get structured transactions back. `?format=json` (default) returns ParseResponse; `?format=csv` returns the engine-serialized CSV. `redact=true` returns the redacted document bytes. The file is parsed in memory and never written to disk.",
        "operationId": "parse_v1_parse_post",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "csv"
              ],
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_v1_parse_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BankInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "BankInfo"
      },
      "BanksResponse": {
        "properties": {
          "banks": {
            "items": {
              "$ref": "#/components/schemas/BankInfo"
            },
            "type": "array",
            "title": "Banks"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          }
        },
        "type": "object",
        "required": [
          "banks",
          "engine_version"
        ],
        "title": "BanksResponse"
      },
      "Body_parse_v1_parse_post": {
        "properties": {
          "pdf": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Pdf"
          },
          "bank": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank"
          },
          "redact": {
            "type": "boolean",
            "title": "Redact",
            "default": false
          },
          "turnstile_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "type": "object",
        "required": [
          "pdf"
        ],
        "title": "Body_parse_v1_parse_post"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "error_class": {
            "type": "string",
            "title": "Error Class"
          },
          "format_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Version"
          },
          "marker_coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marker Coverage"
          }
        },
        "type": "object",
        "required": [
          "error",
          "error_class"
        ],
        "title": "ErrorResponse",
        "description": "Error envelope for all non-2xx responses. `error_class` is one of:\n\n- `EncryptedSourceError`: source is password-protected\n- `EmptyStatementError`: parsed clean, zero rows (see `marker_coverage`)\n- `LayoutDriftError`: bank detected, structure broke\n- `ReconciliationError`: totals don't match the statement\n- `ParseError`: last-resort parse failure\n- `AuthError` / `PayloadTooLarge` / `RateLimitError` / `ServiceUnavailable` /\n  `WorkerError`: framework concerns\n\n`marker_coverage` is populated only for `EmptyStatementError`: high coverage with\nzero rows \u2248 legitimately empty; lower coverage \u2248 silent layout drift."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "env": {
            "type": "string",
            "enum": [
              "live",
              "test"
            ],
            "title": "Env",
            "default": "test"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "KeyCreateRequest"
      },
      "KeyCreatedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "env": {
            "type": "string",
            "title": "Env"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "prefix",
          "name",
          "env",
          "tier"
        ],
        "title": "KeyCreatedResponse",
        "description": "Returned once, on creation. The only time the raw key is ever exposed."
      },
...
      "KeyListResponse": {
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/KeyInfo"
            },
            "type": "array",
            "title": "Keys"
          }
        },
        "type": "object",
        "required": [
          "keys"
        ],
        "title": "KeyListResponse"
      },
      "ParseResponse": {
        "properties": {
          "format_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Version"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatementMetadataOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsOut"
          },
          "row_wise_reconcilable": {
            "type": "boolean",
            "title": "Row Wise Reconcilable"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "type": "array",
            "title": "Transactions"
          }
        },
        "type": "object",
        "required": [
          "format_version",
          "metadata",
          "totals",
          "row_wise_reconcilable",
          "transactions"
        ],
        "title": "ParseResponse"
      },
...
      "UsageResponse": {
        "properties": {
          "api_key_id": {
            "type": "string",
            "title": "Api Key Id"
          },
          "period_parses": {
            "type": "integer",
            "title": "Period Parses"
          },
          "projected_invoice_usd": {
            "type": "string",
            "title": "Projected Invoice Usd"
          }
        },
        "type": "object",
        "required": [
          "api_key_id",
          "period_parses",
          "projected_invoice_usd"
        ],
        "title": "UsageResponse"
      },
...
    }
  },
...
}